refactor(speech): extract frame glue from lifecycle
Move push-to-talk handling, transcript injection, notices, and the
unattended-recording guard into app/speech/input.rs so lifecycle.rs
stays under the line cap. No behavior change.
